Buy Process working again
changed pixel size of character items, fixed bugs in the buy process

+++ Assets/ScriptableObjects/WearableItemScriptableObject.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class WearableItemScriptableObject : ScriptableObject
 {
     public int id;
-    public int Price;
+    public float Price;
     public Sprite ItemImage;
     public bool IsUnlocked = false;
     public WearableItem ItemType;
